refactor: Move AppMain to App root
